Comments reolved -code refactoring
