+++ src/controllers/todoController.ts
@@ -3,15 +3,9 @@ import { Request, Response } from "express";
 import expressAsyncHandler from "express-async-handler";
 
 export const getTodos = expressAsyncHandler(
-  async (req: Request, res: Response) => {
+  async (req: Request, res: Response, next) => {
     const request: any = req;
     const todos: any = await Todo.find({ user: request.user._id });
-
-    if (todos.length == 0) {
-      res.status(404);
-      throw new Error("Todos not Found");
-    }
-
     res.status(200).json(todos);
   }
 );
